model_training_function/requirements.txt
functions-framework
pandas
google-cloud-bigquery
google-cloud-storage
pyarrow
prophet
joblib